+++ core/src/main/java/io/questdb/cutlass/line/tcp/TableUpdateDetails.java
@@ -48,7 +48,7 @@ public class TableUpdateDetails implements Closeable {
     private int writerThreadId;
     // Number of rows processed since the last reshuffle, this is an estimate because it is incremented by
     // multiple threads without synchronisation
-    private int eventsProcessedSinceReshuffle = 0;
+    private long eventsProcessedSinceReshuffle = 0;
     private TableWriter writer;
     private boolean assignedToJob = false;
     private long lastMeasurementMillis = Long.MAX_VALUE;
@@ -127,7 +127,7 @@ public class TableUpdateDetails implements Closeable {
         }
     }
 
-    public int getEventsProcessedSinceReshuffle() {
+    public long getEventsProcessedSinceReshuffle() {
         return eventsProcessedSinceReshuffle;
     }
